# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BIZRA Node0 Genesis API
  description: |
    # BIZRA Node0 Genesis - Enterprise AI Platform API

    The BIZRA Node0 Genesis API provides a comprehensive interface for interacting with
    the enterprise AI platform. This API enables:

    - **AI Query Processing**: Submit queries to LLM backends (Ollama, LM Studio)
    - **User Management**: Authentication, authorization, and session management
    - **Knowledge Graph**: Interact with the Neo4j-based knowledge graph
    - **Vector Search**: Semantic search using Qdrant vector database
    - **System Monitoring**: Health checks, metrics, and status endpoints

    ## Authentication

    All API endpoints (except health and public endpoints) require authentication via:
    - Bearer token in the `Authorization` header
    - API key in the `X-API-Key` header

    ## Rate Limiting

    API requests are rate-limited to ensure fair usage:
    - Standard tier: 100 requests/minute
    - Premium tier: 1000 requests/minute
    - Enterprise tier: Unlimited (subject to fair use policy)

    ## Versioning

    The API is versioned via URL path (`/api/v1`). Breaking changes will result in a new version.
    Deprecated endpoints will be supported for at least 6 months after deprecation notice.

  version: 1.0.0
  contact:
    name: BIZRA Engineering Team
    email: engineering@bizra.io
    url: https://bizra.io
  license:
    name: MIT
...
  x-logo:
...
    altText: BIZRA Logo
servers:
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate a user and receive access tokens
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate the current session and revoke tokens
      operationId: logoutViaDelete
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: unauthorized
            message: Authentication required
            code: 401
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        mfa_code:
          type: string
          pattern: '^\d{6}$'
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
          description: Token expiry in seconds
        user:
          $ref: '#/components/schemas/User'
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [user, admin, superadmin]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...